+++ packages/SwingSet/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.16.2](https://github.com/Agoric/agoric-sdk/compare/@agoric/swingset-vat@0.16.1...@agoric/swingset-vat@0.16.2) (2021-04-16)
+
+**Note:** Version bump only for package @agoric/swingset-vat
+
+
+
+
+
 ## [0.16.1](https://github.com/Agoric/agoric-sdk/compare/@agoric/swingset-vat@0.16.0...@agoric/swingset-vat@0.16.1) (2021-04-14)
 
 **Note:** Version bump only for package @agoric/swingset-vat