+++ packages/spawner/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.4.9](https://github.com/Agoric/agoric-sdk/compare/@agoric/spawner@0.4.8...@agoric/spawner@0.4.9) (2021-04-16)
+
+**Note:** Version bump only for package @agoric/spawner
+
+
+
+
+
 ## [0.4.8](https://github.com/Agoric/agoric-sdk/compare/@agoric/spawner@0.4.7...@agoric/spawner@0.4.8) (2021-04-14)
 
 **Note:** Version bump only for package @agoric/spawner