+++ packages/swingset-runner/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.12.11](https://github.com/Agoric/agoric-sdk/compare/@agoric/swingset-runner@0.12.10...@agoric/swingset-runner@0.12.11) (2021-04-16)
+
+**Note:** Version bump only for package @agoric/swingset-runner
+
+
+
+
+
 ## [0.12.10](https://github.com/Agoric/agoric-sdk/compare/@agoric/swingset-runner@0.12.9...@agoric/swingset-runner@0.12.10) (2021-04-14)
 
 **Note:** Version bump only for package @agoric/swingset-runner