+++ packages/transform-metering/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.4.7](https://github.com/Agoric/agoric-sdk/compare/@agoric/transform-metering@1.4.6...@agoric/transform-metering@1.4.7) (2021-04-16)
+
+**Note:** Version bump only for package @agoric/transform-metering
+
+
+
+
+
 ## [1.4.6](https://github.com/Agoric/agoric-sdk/compare/@agoric/transform-metering@1.4.5...@agoric/transform-metering@1.4.6) (2021-04-07)
 
 **Note:** Version bump only for package @agoric/transform-metering