+++ packages/ui-components/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.4](https://github.com/Agoric/agoric-sdk/compare/@agoric/ui-components@0.2.3...@agoric/ui-components@0.2.4) (2021-04-16)
+
+**Note:** Version bump only for package @agoric/ui-components
+
+
+
+
+
 ## [0.2.3](https://github.com/Agoric/agoric-sdk/compare/@agoric/ui-components@0.2.2...@agoric/ui-components@0.2.3) (2021-04-14)
 
 **Note:** Version bump only for package @agoric/ui-components